Enhance file upload functionality and related services
- Added methods to `IDirectoryProvider` for folder management.
- Changed `AmountPaid` type in `PaymentTransaction` to `double`.
- Introduced `FileUploadStatus` enum for upload states.
- Added `DbSet<FileUpload>` to `EzioHostDbContext`.
- Updated `UserController` to include user ID in operations.
- Modified `VideoController` to utilize `IDirectoryProvider`.
- Updated project to target x64 and added `BlazorBlobStream`.
- Enhanced `Program.cs` with new service registrations.
- Implemented `DirectoryProvider` for folder management.
- Deleted unused binary file `30s.mp4`.
- Improved `UploadFile.razor` for better upload handling.
- Created `FileUpload` entity with repository and service.
- Added `UploadController` for file upload management.
- Introduced `SeekableFileStream` for reading file slices.
- Updated `UploadInfoDto` for file upload metadata.
- Adjusted `AuthController` for user authentication.
- Updated navigation menu to include upload feature.
- Enhanced JavaScript interop for file slice reading.

+++ Back-end/EzioHost.Domain/Entities/PaymentTransaction.cs
@@ -14,7 +14,7 @@ namespace EzioHost.Domain.Entities
 
         public SubscriptionPlan? SubscriptionPlan { get; set; }
 
-        public decimal AmountPaid { get; set; }
+        public double AmountPaid { get; set; }
 
         [Required]
         public string PaymentProvider { get; set; } = "PayPal";
